:hammer: moving svg files, updating integration
