+++ gatsby-config.js
@@ -142,7 +142,7 @@ module.exports = {
           resolve: "gatsby-plugin-react-svg",
           options: {
             rule: {
-              include: /svg/ // See below to configure properly
+              include: /svg/, // See below to configure properly
             }
           }
         },
