+++ src/styles/main.scss
@@ -51,10 +51,12 @@ body {
 
 }
 
+/*
 .svg-icon {
     margin:0;
     padding:0 9px 0 0;
 }
+*/
 
 .btn-ojmain {
     color: #fff;
